More CMake cleanup
Made the GKlib apps disabled by default if the GKlib library is not the
top CMake project and enabled if it is. This way, when GKlib is a
dependency of another project, the apps are not build by default. An
option, GKLIB_BUILD_APPS, was added to allow projects depending on GKlib
to also build the apps.

Also renamed the alias target GKlib::gklib -> GKlib::GKlib to be
consistent with the imported target.